[#1087] Refactored CommitDiffPanel
