+++ frontend/src/commit-diff-panel/CommitDiffPanel.less
@@ -1,4 +1,4 @@
-.DiffPanel {
+.CommitDiffPanel {
 
   @borderColor: #eee;
