+++ frontend/src/commit-diff-panel/chunk-table/ChunkTableRow.tsx
