+++ frontend/src/commit-diff-panel/diff/BinaryFileInfo.tsx
@@ -1,3 +1,5 @@
+/// <reference path='../../common/Diff.d.ts' />
+
 import * as React from 'react';
 
 interface BinaryFileInfoProps {
